+++ DataManager/ClientApp/minerva/ts/components/product-layout/product-navigator.component.tsx
@@ -1,24 +1,41 @@
 import { Fragment, useEffect, useState } from 'react';
 import { Link, useParams, useLocation } from 'react-router';
 import { Rosie } from 'rosie/core';
-
-import { NavItem, ProductNavigatorModel } from 'minerva/core';
+import { NavItem, ProductDashboardTreeModel } from 'minerva/core';
 
 export function ProductNavigator() {
   const params = useParams(),
         [navigation, setNavigation] = useState([] as NavItem[]);
 
   useEffect(() => {
-    const navigator$ = ProductNavigatorModel.subscribe(setNavigation);
-    return () => { navigator$.unsubscribe(); }
-  }, []);
-
-  useEffect(() => {
     const { productId } = params;
     if (productId) {
-      ProductNavigatorModel.load({ pathParams: { productId } });
+      loadDashboardTree(productId);
     }
   }, [params])
+
+  async function loadDashboardTree(productId: string) {
+    const dashboardTree = (await ProductDashboardTreeModel.fetch({ pathParams: { productId } })) ?? [],
+          navigator: NavItem[] = [{
+            navId: 'dashboards',
+            navName: 'Dashboards',
+            children: dashboardTree
+          }, {
+            navId: 'management',
+            navName: 'Management',
+            children: [{
+              navId: 'events',
+              navName: 'Events',
+              navPath: `/products/${productId}/events`
+            }, {
+              navId: 'settings',
+              navName: 'Settings',
+              navPath: `/products/${productId}/settings`
+            }]
+          }];
+
+    setNavigation(navigator);
+  }
 
   return <>
     <NavItemList items={navigation} level={0} />
@@ -31,9 +48,9 @@ function NavItemList(props: { items: NavItem[], level: number }) {
 
   return <>
     {items.map(navItem => {
-      if (navItem.children && navItem.children.length > 0) {
+      if (!navItem.navPath || (navItem.children && navItem.children.length > 0)) {
         return <Fragment key={navItem.navId}>
-          <div className="nav-link disabled text-body-tertiary py-1 pe-1" style={{paddingLeft: 4 + 8 * level}}>{navItem.navName}</div>
+          <div className="nav-link disabled py-1 pe-1" style={{paddingLeft: 4 + 8 * level}}>{navItem.navName}</div>
           <NavItemList items={navItem.children} level={level + 1} />
         </Fragment>
       }